:art: refactor: Break out the collaborators into separate modules
A 1,000 line long test module is not cool.
